fix delete function

+++ src/app/features/staff-list/components/edit-staff/edit-staff.component.ts
@@ -17,8 +17,10 @@ import { SelectComponent } from '../../../../shared/ui-components/atoms/select/s
 import { StaffAccount } from '../../models/staff';
 import { StaffService } from '../../services/staff.service';
 import { MatMenuModule } from '@angular/material/menu';
-import { MatDialog } from "@angular/material/dialog";
-import { UpdateEmailDialogComponent } from '../change-email-dialog/change-email-dialog';
+import { MatDialog } from '@angular/material/dialog';
+import { UpdateEmailDialogComponent } from '../../../../shared/ui-components/templates/change-email-dialog/change-email-dialog';
+import { ConfirmDeleteComponent } from '../../../../shared/ui-components/templates/confirm-delete/confirm-delete.component';
+import { ChangePasswordDialogComponent } from '../../../../shared/ui-components/templates/change-password-dialog/change-password-dialog.component';
 
 @Component({
   selector: 'app-edit-staff',
@@ -34,7 +36,6 @@ import { UpdateEmailDialogComponent } from '../change-email-dialog/change-email-
     SelectComponent,
     AsyncPipe,
     MatMenuModule,
-            
   ],
   templateUrl: './edit-staff.component.html',
   styleUrl: './edit-staff.component.scss',
@@ -51,6 +52,7 @@ export class EditStaffComponent {
   lookupService = inject(LookupService);
   id = input.required<string>();
   cloneStaff = new StaffAccount('Receptionist');
+
   constructor() {}
   ngOnInit(): void {
     this.staffService.getStaff(this.id()).subscribe((res) => {
@@ -89,32 +91,59 @@ export class EditStaffComponent {
       });
   }
   deleteStaff() {
-    this.cloneStaff.isDeleted = true;
-    this.staffService.deleteStaff(this.cloneStaff.id).subscribe((res) => {
-      if (!res.error) {
-        this.snackbarService.success('DELETE_STAFF_SUCCESS');
-        this.router.navigate(['/', APP_ROUTES.STAFF_LIST]);
+    this.dialog
+      .open(ConfirmDeleteComponent, {
+        data: {
+          translationTemplate: this.translationTemplate,
+          headerText: 'DELETE_STAFF_HEADER',
+          content: 'DELETE_STAFF_CONTENT',
+          cancelText: 'NO',
+          okText: 'YES',
+        },
+      })
+      .afterClosed()
+      .subscribe((res) => {
+        if (res) {
+          this.cloneStaff.isDeleted = true;
+          this.staffService.deleteStaff(this.cloneStaff.id).subscribe((res) => {
+            if (!res.error) {
+              this.snackbarService.success('DELETE_STAFF_SUCCESS');
+              this.router.navigate(['/', APP_ROUTES.STAFF_LIST]);
+            }
+          });
+        }
+      });
+  }
+  openChangeEmailDialog() {
+    const ref = this.dialog.open(UpdateEmailDialogComponent, {
+      data: {
+        userId: this.staffAccount().id,
+        email: this.staffAccount().email,
+      },
+      width: '400px',
+      minWidth: '300px',
+    });
+
+    ref.afterClosed().subscribe((result) => {
+      if (result?.success) {
+        // Update your local signal so the template shows the new email
+        this.staffAccount.update((acc) => ({ ...acc, email: result.email }));
+        this.snackbarService.success('CHANGE_EMAIL_SUCCESS');
       }
     });
   }
-  openChangeEmailDialog() {
-  const ref = this.dialog.open(UpdateEmailDialogComponent, {
-    data: {
-      userId: this.staffAccount().id,
-      email:  this.staffAccount().email
-    },
-    width: '400px',
-    minWidth: '300px'
-  });
-
-  ref.afterClosed().subscribe(result => {
-    if (result?.success) {
-      // Update your local signal so the template shows the new email
-      this.staffAccount.update(acc => ({ ...acc, email: result.email }));
-      this.snackbarService.success('CHANGE_EMAIL_SUCCESS');
-    }
-  });
-}
-
-
+  openChangePasswordDialog() {
+    const ref = this.dialog.open(ChangePasswordDialogComponent, {
+      data: {
+        userId: this.staffAccount().id,
+      },
+      width: '400px',
+      minWidth: '300px',
+    });
+    ref.afterClosed().subscribe((result) => {
+      if (result?.success) {
+        this.snackbarService.success('CHANGE_PASSWORD_SUCCESS');
+      }
+    });
+  }
 }
